Port bink_playback.obj: lift 5 remaining functions (bink_playback.obj) (88.6% VC71, 10/10 equiv)
Functions ported:
- bink_memory_callback_alloc @ 0x1c5ab0 (bink_playback.obj) [87.2% VC71, 137/145 insns]
- FUN_001c6900 @ 0x1c6900 (bink_playback.obj) [77.0% VC71, 171/198 insns]
- FUN_001c6b20 @ 0x1c6b20 (bink_playback.obj) [92.0% VC71, 92/84 insns]
- FUN_001c6bf0 @ 0x1c6bf0 (bink_playback.obj) [100.0% VC71, 1/1 insns]
- FUN_001c6c00 @ 0x1c6c00 (bink_playback.obj) [86.7% VC71, 58/62 insns]

Callee decls corrected (signature only, name unchanged):
- FUN_001c6d90 @ 0x1c6d90 (game_sound.obj)
    void FUN_001c6d90(void)
 -> bool FUN_001c6d90(file_ref_t *file, void *format_out)

kb_meta.json updates:
- 11 symbol(s) updated

Coverage: 37.1% (3029/8170 symbols)

kb.json object correction: 6 functions moved out of bink_playback.obj
into saved_game_files.obj (0x1c4da0, 0x1c4f30, 0x1c5010, 0x1c53f0,
0x1c5560, 0x1c58f0). They are not bink code and were left unported:
0x1c4da0/0x1c53f0/0x1c5560 push the saved_game_files.c __FILE__ literal
at 0x2ba8e8 into display_assert; 0x1c4f30 and 0x1c5010 carry no own
__FILE__ but are contiguous with that block, 0x1c5010 is called only by
the two proven ones, and 0x1c4f30 owns the "XDeleteSaveGame() failed to
delete profile" string -- neither touches a bink global. 0x1c58f0 is not
source at all: E9 1B F7 FF FF (JMP 0x1c5010) plus 0x90 padding, an MSVC
incremental-link thunk, filed with its jump target. Evidence recorded
per address in kb_meta.json. bink_playback.obj 31 -> 25 functions,
saved_game_files.obj 56 -> 62; 8014 addrs, 0 duplicates.

Behavioural verification (all three sub-90% functions, required because
byte-match alone is weak evidence there):
- bink_memory_callback_alloc: 240/240 seeds over 8 branch families,
  87.9% union coverage = 100% of reachable code, 3 distinct returns,
  0 mem-trace divergences, stub-arg differential armed and clean.
- FUN_001c6900: 180/180 over 9 paths, 100% union coverage, all three
  sample-rate arms selected by data through a real memcmp, both success
  conditions and every rejection path. A negative control (swapping the
  0x5622/0xac44 stores) fails 5/5, so the lane detects this bug class.
- FUN_001c6c00: 140/140 over 7 paths, 95.9% union = 100% of reachable
  code. Caveat: the call-sequence differential is disarmed for this
  target (same-TU sibling asymmetry -- the per-function oracle stubs
  FUN_001c6880/0x1c6900 while the candidate runs them for real), so its
  evidence is return value + mem-trace + scratch only.

FUN_001c6900's 77.0% is a structural cap, not missing logic: 56 unpaired
reference instructions itemized as two full copies of a 19-insn shared
tail plus duplicated epilogue fragments, and the ref-13-vs-ours-12 call
delta is a duplicated file_close, not a dropped call. The 44100 arm
jumps into the 11025 arm's tail at 0x1c6ae8, proving one shared tail in
the original source.

6 targets left unported and parked: the 5 saved_game_files.c functions
above (correct home is that TU, not this one) and the 0x1c58f0 thunk.

+++ src/halo/input/input_xbox.c
@@ -669,11 +669,11 @@ bool input_has_gamepad(int16_t gamepad_index)
   return input_gamepad_handles()[gamepad_index] != 0;
 }
 
-void *input_get_gamepad_state(int gamepad_index)
+void *input_get_gamepad_state(int16_t gamepad_index)
 {
   int16_t index;
 
-  index = (int16_t)gamepad_index;
+  index = gamepad_index;
   assert_halt(index >= 0 && index < MAXIMUM_GAMEPADS);
   if (input_gamepad_handles()[index] != 0) {
     if (*input_suppressed())
